Rminr refactoring
