+++ src/components/AutoGraticule/index.tsx
@@ -1,55 +1,32 @@
-import { LatLngBounds, LatLngExpression, LayerOptions, LeafletEventHandlerFnMap, PolylineOptions, Map, LayerGroup, Util, Polyline, LatLng, marker, divIcon, latLngBounds, Control } from 'leaflet';
+import {
+    LatLngBounds,
+    LatLngExpression,
+    LayerOptions,
+    LeafletEventHandlerFnMap,
+    PolylineOptions,
+    Map,
+    LayerGroup,
+    Util,
+    Polyline,
+    LatLng,
+    marker,
+    divIcon,
+    latLngBounds,
+} from 'leaflet';
 import './index.css';
 import { useMap } from 'react-leaflet';
-import { useState, useEffect, useCallback, useRef } from 'react';
+import { useState, useEffect } from 'react';
 import { formatCoordinate } from '../../helpers/formatCoordinate';
-import { useAppSelector } from '../../app/hooks';
-import { selectBounds } from '../../features/geoFeatures/geoFeaturesSlice';
 
-export interface AutoGraticuleOptions extends LayerOptions {
+interface GraticuleOptions extends LayerOptions {
     redraw: keyof LeafletEventHandlerFnMap,
 
     /** Minimum distance between two lines in pixels */
     minDistance: number
 
-    formatter?: (lat: number, isLatitude: boolean) => string
+    // method that will format the lat/long value
+    formatter?: (value: number, isLatitude: boolean) => string
 }
-
-/** helper component provides the map graticule */
-export const HomeControl: React.FC = () => {
-    const map = useMap()
-    const currentBounds = selectBounds(useAppSelector(state => state.featureCollection))
-    const [existing, setExisting] = useState<Control | undefined>(undefined)
-    const controlInitialised = useRef(false); 
-
-    const doHome = useCallback(() => {
-        map.flyToBounds(currentBounds)
-    }, [map, currentBounds])
-
-    // note: we need to return a Leaflet control.  It will have a house
-    // icon that when clicked will fly the map to the current bounds, via the DoHome function
-    useEffect(() => {
-        if (map && !controlInitialised.current) {
-            if (existing) {
-                existing.remove()
-            }
-            const control = new Control({position: 'topleft'})
-            control.onAdd = () => {
-                const div = document.createElement('div')
-                div.className = 'leaflet-bar leaflet-control leaflet-control-custom'
-                div.innerHTML = '<a href="#" title="Home" role="button" aria-label="Home" class="leaflet-bar-part leaflet-bar-part-single" id="home" onclick="return false;">🏠</a>'
-                div.onclick = doHome
-                return div
-            }
-            control.addTo(map)
-            setExisting(control)
-            controlInitialised.current = true
-        }
-    },[map])
-
-    return null
- }
-
 
 /** helper component provides the map graticule */
 export const Graticule: React.FC = () => {
@@ -63,7 +40,7 @@ export const Graticule: React.FC = () => {
       const formatter = (value: number, isLat: boolean): string => {
         return formatCoordinate(value, isLat)
       }
-      const options: AutoGraticuleOptions = {
+      const options: GraticuleOptions = {
         redraw: 'moveend',
         minDistance: 100,  // Minimum distance between two lines in pixels
         formatter: formatter
@@ -78,7 +55,7 @@ export const Graticule: React.FC = () => {
 
 export default class AutoGraticule extends LayerGroup {
 
-    options: AutoGraticuleOptions = {
+    options: GraticuleOptions = {
         redraw: 'moveend',
         minDistance: 100,  // Minimum distance between two lines in pixels
         formatter: undefined
@@ -95,7 +72,7 @@ export default class AutoGraticule extends LayerGroup {
     _bounds!: LatLngBounds;
 
 
-    constructor(options?: Partial<AutoGraticuleOptions>) {
+    constructor(options?: Partial<GraticuleOptions>) {
         super();
         Util.setOptions(this, options);
     }
